+++ exec_graph/include/exec_graph/graph/graph_document.hpp
@@ -16,6 +16,7 @@ struct GraphEdge {
 };
 
 struct GraphDocument {
+    std::string working_directory;
     std::vector<GraphNode> nodes;
     std::vector<GraphEdge> edges;
 };